+++ lib/icon_content.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+// カードの中身（性別選択）
+class IconContent extends StatelessWidget {
+
+  IconContent({@required this.icon, @required this.label});
+
+  final IconData icon;
+  final String label;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: <Widget>[
+        Icon(
+          icon,
+          size: 80.0,
+        ),
+        SizedBox(
+          height: 15.0,
+        ),
+        Text(
+          label,
+          style: TextStyle(
+            fontSize: 18.0,
+            color: Color(0xFF8D8E98),
+          ),
+        ),
+      ],
+    );
+  }
+}
