+++ lib/reusable_card.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+// コンテナ1つ分
+class ReusableCard extends StatelessWidget {
+  ReusableCard({@required this.colour, @required this.cardChild});
+
+  // final：プログラム開始後のある時点で一回だけ初期化される。再代入不可。
+  final Color colour;
+  final Widget cardChild;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: cardChild,
+      margin: EdgeInsets.all(15.0),
+      decoration: BoxDecoration(
+        color: colour,
+        borderRadius: BorderRadius.circular(10.0),
+      ),
+    );
+  }
+}
